Sar2016

+++ src/husacct/define/domain/services/ModuleDomainService.java
@@ -136,8 +136,20 @@ public class ModuleDomainService {
 		return SoftwareArchitecture.getInstance().getModuleById(moduleId);
 	}
 
+	// Returns the ModuleStrategy of the found module, or throws an exception if no module is found.
 	public ModuleStrategy getModuleByLogicalPath(String logicalPath) {
 		return SoftwareArchitecture.getInstance().getModuleByLogicalPath(logicalPath);
+	}
+
+	// Returns the ModuleStrategy of the found module, or null if no module is found.
+	public ModuleStrategy getModuleByLogicalPath_NoException(String logicalPath) {
+		ModuleStrategy foundModule;
+		try {
+			foundModule = SoftwareArchitecture.getInstance().getModuleByLogicalPath(logicalPath);
+	    } catch (Exception e) {
+	    	foundModule = null;
+	    }
+	    return foundModule;
 	}
 
 	public ModuleStrategy getModuleIdBySoftwareUnit(SoftwareUnitDefinition su) {
@@ -147,7 +159,10 @@ public class ModuleDomainService {
 	public String getModuleNameById(long moduleId) {
 		String moduleName = new String();
 		if (moduleId != -1) {
-			moduleName = SoftwareArchitecture.getInstance().getModuleById(moduleId).getName();
+			ModuleStrategy module = SoftwareArchitecture.getInstance().getModuleById(moduleId);
+			if (module != null) {
+				moduleName = module.getName();
+			}
 		}
 		return moduleName;
 	}
@@ -183,10 +198,12 @@ public class ModuleDomainService {
 	public ArrayList<Long> getSiblingModuleIds(long moduleId) {
 		ArrayList<Long> childModuleIdList = new ArrayList<Long>();
 		if (moduleId != -1) {
-			long parentModuleId = SoftwareArchitecture.getInstance().getParentModuleIdByChildId(moduleId);
-			childModuleIdList = getSubModuleIds(parentModuleId);
 			ModuleStrategy module = SoftwareArchitecture.getInstance().getModuleById(moduleId);
-			childModuleIdList.remove(module.getId());
+			if (module != null) {
+				long parentModuleId = SoftwareArchitecture.getInstance().getParentModuleIdByChildId(moduleId);
+				childModuleIdList = getSubModuleIds(parentModuleId);
+				childModuleIdList.remove(module.getId());
+			}
 		}
 		return childModuleIdList;
 	}
@@ -204,11 +221,13 @@ public class ModuleDomainService {
 		ArrayList<Long> childModuleIdList = new ArrayList<Long>();
 		if (parentModuleId != -1) {
 			ModuleStrategy parentModule = SoftwareArchitecture.getInstance().getModuleById(parentModuleId);
-			for (ModuleStrategy module : parentModule.getSubModules()) {
-				childModuleIdList.add(module.getId());
-				ArrayList<Long> subModuleIdList = getSubModuleIds(module.getId());
-				for (Long l : subModuleIdList) {
-					childModuleIdList.add(l);
+			if (parentModule != null) {
+				for (ModuleStrategy module : parentModule.getSubModules()) {
+					childModuleIdList.add(module.getId());
+					ArrayList<Long> subModuleIdList = getSubModuleIds(module.getId());
+					for (Long l : subModuleIdList) {
+						childModuleIdList.add(l);
+					}
 				}
 			}
 		} else {
@@ -236,13 +255,15 @@ public class ModuleDomainService {
 
 	public void removeModuleById(long moduleId) {
 		ModuleStrategy module = SoftwareArchitecture.getInstance().getModuleById(moduleId);
-      try{
-		SoftwareArchitecture.getInstance().removeModule(module);
-	    JtreeController.instance().registerTreeRemoval(module);
-      }catch(Exception e) {
-    	  logger.error(e);
-      }
-		ServiceProvider.getInstance().getDefineService().notifyServiceListeners();
+		if (module != null) {
+			try{
+				SoftwareArchitecture.getInstance().removeModule(module);
+				JtreeController.instance().registerTreeRemoval(module);
+			}catch(Exception e) {
+				logger.error(e);
+			}
+			ServiceProvider.getInstance().getDefineService().notifyServiceListeners();
+		}
 	}
 
 	public void sortModuleChildren(ModuleStrategy module) {
@@ -255,16 +276,18 @@ public class ModuleDomainService {
 
 	public void updateModule(long moduleId, String moduleName, String moduleDescription) {
 		ModuleStrategy module = SoftwareArchitecture.getInstance().getModuleById(moduleId);
-		StateService.instance().addUpdateModule(moduleId, new String[] { module.getName(), moduleDescription },
-				new String[] { moduleName, moduleDescription });
-		module.setName(moduleName);
-		module.setDescription(moduleDescription);
-		ServiceProvider.getInstance().getDefineService().notifyServiceListeners();
+		if (module != null) {
+			StateService.instance().addUpdateModule(moduleId, new String[] { module.getName(), moduleDescription },
+					new String[] { moduleName, moduleDescription });
+			module.setName(moduleName);
+			module.setDescription(moduleDescription);
+			ServiceProvider.getInstance().getDefineService().notifyServiceListeners();
+		}
 	}
 	
 	public void updateModuleType(long moduleId, String newType) {
 		ModuleStrategy oldModule = SoftwareArchitecture.getInstance().getModuleById(moduleId);
-		if(oldModule.getId() > 0){
+		if((oldModule != null) && (oldModule.getId() > 0)){
 			if (oldModule.getType() != newType){
 				DefaultRuleDomainService service = new DefaultRuleDomainService();
 				service.removeDefaultRules(oldModule);
@@ -279,11 +302,12 @@ public class ModuleDomainService {
 	
 	public void updateFacade(long moduleId, String moduleName){
 		ModuleStrategy parent = SoftwareArchitecture.getInstance().getModuleById(moduleId);
-		for(ModuleStrategy subModule : parent.getSubModules()){
-			if(subModule.getType().equals("Facade")){
-				subModule.setName("Interface<"+moduleName+">");
+		if (parent != null) {
+			for(ModuleStrategy subModule : parent.getSubModules()){
+				if(subModule.getType().equals("Facade")){
+					subModule.setName("Interface<"+moduleName+">");
+				}
 			}
 		}
 	}
-	
 }
